+++ front/src/app/itinerary/module/moduleEditor/module-editor.component.ts
@@ -139,7 +139,7 @@ export class ModuleEditorComponent implements OnInit {
 
   expandNode(node: Module) {
     if (node.blocks) {
-      this.actualTab.studentAddOpenedNode(node, this.moduleId);
+      this.actualTab.studentAddOpenedNode(this.module, node.id);
       console.log(this.actualTab.openedModuleNodes);
     } else {
       this.actualTab.studentAddOpenedLesson(this.module, node.id);